requirements_mvp.txt
# MVP Production requirements (Azure + LangSmith + OpenAI)
openai>=1.30.0
langsmith>=0.1.36
langchain>=0.2.10
langgraph>=0.2.14
pydantic>=2.7.0
fastapi>=0.111.0
uvicorn>=0.30.0
azure-servicebus>=7.12.1
azure-identity>=1.17.1
azure-cosmos>=4.7.0
httpx>=0.27.0
requests>=2.31.0
rich>=13.7.1
python-dotenv>=1.0.1
pytest>=8.2.1
pytest-asyncio>=0.23.7
pinecone>=5.0.0
redis>=5.0.0
